Remove PlayExistingCommand
the query logic leads to endless queries when editing existing
aggregates and the cmd button is placed in the bottom bar.
In that case, the state query is already fired in the state view and the
dialog fires it again. The staleTime config in the cmd dialog seems to
be ignored because the state view fires the same query without stale
time.